+++ src/SBInterpolatedImage.cpp
@@ -1,3 +1,23 @@
+// -*- c++ -*-
+/*
+ * Copyright 2012, 2013 The GalSim developers:
+ * https://github.com/GalSim-developers
+ *
+ * This file is part of GalSim: The modular galaxy image simulation toolkit.
+ *
+ * GalSim is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GalSim is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with GalSim.  If not, see <http://www.gnu.org/licenses/>
+ */
 
 //#define DEBUGLOGGING
 
@@ -458,7 +478,7 @@ namespace galsim {
         }
         dbg<<"Done: flux = "<<flux<<std::endl;
         // Should have added up to the total flux.
-        assert( std::abs(flux - fluxTot) < 1.e-6 * fluxTot );
+        assert( std::abs(flux - fluxTot) < 1.e-6 * std::abs(fluxTot) );
         if (d1 == 0) {
             dbg<<"No smaller radius found.  Keep current value of stepk\n";
             return;