+++ tests/lensing_reference_data/shearfield_ref.py
@@ -1,8 +1,27 @@
+# Copyright 2012, 2013 The GalSim developers:
+# https://github.com/GalSim-developers
+#
+# This file is part of GalSim: The modular galaxy image simulation toolkit.
+#
+# GalSim is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# GalSim is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with GalSim.  If not, see <http://www.gnu.org/licenses/>
+#
 import numpy as np
 import galsim
 
 # The reference data for this test was generated with this script, using the version of the code on
-# branch #291 at commit 7a1852ede4480f5b366c5f8ac0fe3fa01b9d39dc
+# branch #305 at commit e06b7604090cab8de6ec681cca8b74ace386d240, i.e.,
+# https://github.com/GalSim-developers/GalSim/commit/e06b7604090cab8de6ec681cca8b74ace386d240
 
 # random seed, etc.
 outfile = 'shearfield_reference.dat'
@@ -15,7 +34,7 @@ dx = 1.
 # define power spectrum
 ps = galsim.lensing.PowerSpectrum(e_power_function="k**2", b_power_function="k")
 # get shears
-g1, g2 = ps.getShear(grid_spacing = dx, grid_nx = n, rng=rng)
+g1, g2 = ps.buildGriddedShears(grid_spacing = dx, ngrid = n, rng=rng)
 
 # write to file
 g1vec = g1.reshape(n*n)