+++ CHANGELOG.md
@@ -54,6 +54,13 @@ Changes from v0.2 to v0.3
   wish to take some arbitrary input image and manipulate it (shear, convolved, etc.) can use this
   base class as a way of easily carrying out those operations.  (Issue #333)
 
+* There is a new class structure for representing 2D correlation functions, used to describe 
+  correlated noise in images (for example).  The intended user interface is using the
+  `cf = galsim.ImageCorrFunc(image)` constructor, which calculates the discrete correlation function
+  between pixels in the input image.  The `cf` instance acts as a container for an internal 
+  `GSObject` representation of this correlation function, making many `GSObject` operations
+  available (Issue #297).
+
 * Option for shear power spectra: use a tabulated P(k), either input as arrays or read in from a
   file, for example from a cosmological shear power spectrum calculator.  This work also involved
   making a python interface to C++ tables that can be used for interpolation in a more general